requirements.txt
Flask
python-dotenv
firebase-admin
flask-cors
Flask-RESTful
requests 
litellm
openai